+++ flask_rest_api/__init__.py
@@ -7,7 +7,7 @@ from .blueprint import Blueprint  # noqa
 from .pagination import Page  # noqa
 from .error_handler import ErrorHandlerMixin
 
-__version__ = '0.13.1'
+__version__ = '0.14.0'
 
 
 class Api(APISpecMixin, ErrorHandlerMixin):
@@ -41,7 +41,7 @@ class Api(APISpecMixin, ErrorHandlerMixin):
         self._app = app
         self.spec = None
         # Use lists to enforce order
-        self._definitions = []
+        self._schemas = []
         self._fields = []
         self._converters = []
         if app is not None: