+++ flask_rest_api/pagination.py
@@ -18,7 +18,7 @@ from flask import request, current_app
 import marshmallow as ma
 from webargs.flaskparser import FlaskParser
 
-from .utils import get_appcontext
+from .utils import unpack_tuple_response
 from .compat import MARSHMALLOW_VERSION_MAJOR
 
 
@@ -135,6 +135,10 @@ class PaginationMixin:
 
         Otherwise, pagination is handled in the view function.
 
+        The decorated function may return a tuple including status and/or
+        headers, like a typical flask view function. It may not return a
+        ``Response`` object.
+
         See :doc:`Pagination <pagination>`.
         """
         if page is None:
@@ -166,7 +170,8 @@ class PaginationMixin:
                     kwargs['pagination_parameters'] = page_params
 
                 # Execute decorated function
-                result = func(*args, **kwargs)
+                result, status, headers = unpack_tuple_response(
+                    func(*args, **kwargs))
 
                 # Post pagination: use pager class to paginate the result
                 if pager is not None:
@@ -182,10 +187,12 @@ class PaginationMixin:
                         page_header = self._make_pagination_header(
                             page_params.page, page_params.page_size,
                             page_params.item_count)
-                        get_appcontext()['headers'][
+                        if headers is None:
+                            headers = {}
+                        headers[
                             self.PAGINATION_HEADER_FIELD_NAME] = page_header
 
-                return result
+                return result, status, headers
 
             return wrapper