+++ flask_rest_api/response.py
@@ -2,9 +2,13 @@
 
 from functools import wraps
 
+from werkzeug import BaseResponse
 from flask import jsonify
 
-from .utils import deepupdate, get_appcontext
+from .utils import (
+    deepupdate, get_appcontext,
+    unpack_tuple_response, set_status_and_headers_in_response
+)
 from .compat import MARSHMALLOW_VERSION_MAJOR
 
 
@@ -16,8 +20,17 @@ class ResponseMixin:
 
         :param schema: :class:`Schema <marshmallow.Schema>` class or instance.
             If not None, will be used to serialize response data.
-        :param int code: HTTP status code (default: 200).
+        :param int code: HTTP status code (default: 200). Used if none is
+            returned from the view function.
         :param str descripton: Description of the response.
+
+        The decorated function is expected to return the same types of value
+        than a typical flask view function, except the body part may be an
+        object or a list of objects to serialize with the schema, rather than
+        a ``string``.
+
+        If the decorated function returns a ``Response`` object, the ``schema``
+        and ``code`` parameters are only used to document the resource.
 
         See :doc:`Response <response>`.
         """
@@ -37,7 +50,14 @@ class ResponseMixin:
             def wrapper(*args, **kwargs):
 
                 # Execute decorated function
-                result_raw = func(*args, **kwargs)
+                result_raw, status, headers = unpack_tuple_response(
+                    func(*args, **kwargs))
+
+                # If return value is a werkzeug BaseResponse, return it
+                if isinstance(result_raw, BaseResponse):
+                    set_status_and_headers_in_response(
+                        result_raw, status, headers)
+                    return result_raw
 
                 # Dump result with schema if specified
                 if schema is None:
@@ -48,13 +68,15 @@ class ResponseMixin:
                         result_dump = result_dump[0]
 
                 # Store result in appcontext (may be used for ETag computation)
-                get_appcontext()['result_raw'] = result_raw
-                get_appcontext()['result_dump'] = result_dump
+                appcontext = get_appcontext()
+                appcontext['result_raw'] = result_raw
+                appcontext['result_dump'] = result_dump
 
                 # Build response
                 resp = jsonify(self._prepare_response_content(result_dump))
-                resp.headers.extend(get_appcontext()['headers'])
-                resp.status_code = code
+                set_status_and_headers_in_response(resp, status, headers)
+                if status is None:
+                    resp.status_code = code
 
                 return resp