+++ flask_rest_api/utils.py
@@ -1,8 +1,8 @@
 """Utils"""
 
-from collections import defaultdict
 from collections.abc import Mapping
 
+from werkzeug.datastructures import Headers
 from flask import _app_ctx_stack
 from apispec.utils import trim_docstring, dedent
 
@@ -25,10 +25,10 @@ def deepupdate(original, update):
 def get_appcontext():
     """Return extension section from top of appcontext stack"""
 
-    # http://flask.pocoo.org/docs/0.12/extensiondev/#the-extension-code
+    # http://flask.pocoo.org/docs/latest/extensiondev/#the-extension-code
     ctx = _app_ctx_stack.top
     if not hasattr(ctx, 'flask_rest_api'):
-        ctx.flask_rest_api = defaultdict(dict)
+        ctx.flask_rest_api = {}
     return ctx.flask_rest_api
 
 
@@ -62,3 +62,44 @@ def load_info_from_docstring(docstring):
     if description_lines:
         info['description'] = dedent('\n'.join(description_lines))
     return info
+
+
+# Copied from flask
+def unpack_tuple_response(rv):
+    """Unpack a flask Response tuple"""
+
+    status = headers = None
+
+    # unpack tuple returns
+    if isinstance(rv, tuple):
+        len_rv = len(rv)
+
+        # a 3-tuple is unpacked directly
+        if len_rv == 3:
+            rv, status, headers = rv
+        # decide if a 2-tuple has status or headers
+        elif len_rv == 2:
+            if isinstance(rv[1], (Headers, dict, tuple, list)):
+                rv, headers = rv
+            else:
+                rv, status = rv
+        # other sized tuples are not allowed
+        else:
+            raise TypeError(
+                'The view function did not return a valid response tuple.'
+                ' The tuple must have the form (body, status, headers),'
+                ' (body, status), or (body, headers).'
+            )
+
+    return rv, status, headers
+
+
+def set_status_and_headers_in_response(response, status, headers):
+    """Set status and headers in flask Reponse object"""
+    if headers:
+        response.headers.extend(headers)
+    if status is not None:
+        if isinstance(status, int):
+            response.status_code = status
+        else:
+            response.status = status